@@ -11,32 +11,32 @@
   "type": "library",
   "autoload": {
     "psr-4": {
-      "Pccomponentes\\OpenApiMessagingContext\\": "src/"
+      "PcComponentes\\OpenApiMessagingContext\\": "src/"
     }
   },
   "autoload-dev": {
     "psr-4": {
-      "Pccomponentes\\OpenApiMessagingContext\\Tests\\": "tests/"
+      "PcComponentes\\OpenApiMessagingContext\\Tests\\": "tests/"
     }
   },
   "require": {
     "ext-json": "*",
     "php": "^7.4",
-    "symfony/yaml": "^4.0",
-    "symfony/messenger": "^4.0",
+    "symfony/yaml": "^4.4|^5.0",
+    "symfony/messenger": "^4.4|^5.0",
     "justinrainbow/json-schema": "^5.2",
     "behat/behat": "^3.5",
-    "behat/mink": "@dev",
-    "behat/mink-browserkit-driver": "^1.3",
-    "behat/symfony2-extension": "^2.1",
-    "behat/mink-extension": "^2.3",
     "pccomponentes/ddd": "^2.0",
-    "beberlei/assert": "^3.2"
+    "beberlei/assert": "^3.2",
+    "friends-of-behat/symfony-extension": "^2.0",
+    "friends-of-behat/mink": "^1.8",
+    "friends-of-behat/mink-extension": "^2.4",
+    "friends-of-behat/mink-browserkit-driver": "^1.4"
   },
   "require-dev": {
     "pccomponentes/ganchudo": "^1.0",
     "phpunit/phpunit": "^9.0",
-    "symfony/var-dumper": "^4.0",
+    "symfony/var-dumper": "^4.4|^5.0",
     "pccomponentes/coding-standard": "^1.0"
   }
 }
